+++ src/parser/check__map.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   check__map.c                                       :+:      :+:    :+:   */
+/*   check__map2.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: tkim <tkim@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/05/04 10:13:06 by youkim            #+#    #+#             */
-/*   Updated: 2022/06/10 13:28:48 by tkim             ###   ########.fr       */
+/*   Created: 2022/06/10 13:27:18 by tkim              #+#    #+#             */
+/*   Updated: 2022/06/10 13:29:05 by tkim             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,38 +14,24 @@
 #include "parser.h"
 #include "flags.h"
 
-bool	mapformat__is_player(t_mapfmt this)
+void	raw_map_arr__check_valid(t_string_arr this)
 {
-	return (this == MAPFMT__NORTH || this == MAPFMT__SOUTH
-		|| this == MAPFMT__WEST || this == MAPFMT__EAST);
-}
+	t_ivec						it;
+	t_mapfmt					tile;
+	t_mapformat__validator_f	func;
 
-bool	mapformat__is_valid(t_mapfmt this)
-{
-	return (this == MAPFMT__FLOOR
-		|| this == MAPFMT__WALL
-		|| this == MAPFMT__EMPTY
-		|| mapformat__is_player(this));
-}
-
-bool	mapformat__is_valid_bonus(t_mapfmt this)
-{
-	return (mapformat__is_valid(this)
-		|| this == MAPFMT__DOOR
-		|| this == MAPFMT__SPRITE);
-}
-
-bool	mapformat__is_valid_fluidfill(t_mapfmt this)
-{
-	return (!(this == MAPFMT__WALL
-			|| this == MAPFMT__EMPTY
-			|| this == MAPFMT__FILL));
-}
-
-void	mapformat__assert_valid(
-	t_mapfmt this, t_mapformat__validator_f func)
-{
-	if (!(func(this)))
-		std__panic__value__char(
-			"raw_map_arr__check_valid: invalid tile", this);
+	if (BONUS)
+		func = mapformat__is_valid_bonus;
+	else
+		func = mapformat__is_valid;
+	it.y = -1;
+	while (this[++it.y])
+	{
+		it.x = -1;
+		while (this[it.y][++it.x])
+		{
+			tile = this[it.y][it.x];
+			mapformat__assert_valid(tile, func);
+		}
+	}
 }
